# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API for Icloud",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/documents/stats/relevance": {
            "get": {
                "tags": [
                    "Document Stats"
                ],
                "summary": "Get document relevance statistics",
                "operationId": "7f45be5ecc901c31737045b236de3d8d",
                "responses": {
                    "200": {
                        "description": "Document Relevance Stats retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Document Relevance Stats retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document Relevance Stats not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Document Relevance Stats not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/documents/stats/monthly-approvals": {
            "get": {
                "tags": [
                    "Document Stats"
                ],
                "summary": "Get document monthly approvals statistics",
                "operationId": "af925766fa3c8cddefd3688f8bde0514",
                "responses": {
                    "200": {
                        "description": "Document Monthly Stats retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Document Monthly Stats retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document Monthly Stats not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Document Monthly Stats not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/documents/stats/relevance-with-documents": {
            "get": {
                "tags": [
                    "Document Stats"
                ],
                "summary": "Get document relevance statistics with document details",
                "operationId": "3daea1186702050ced49a0b028a3a674",
                "responses": {
                    "200": {
                        "description": "Document Relevance Stats with Documents retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Document Relevance Stats with Documents retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document Relevance Stats with Documents not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Document Relevance Stats with Documents not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "!IMPORTANT dont forget the prefix Bearer! ej: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}